requirements.txt
beautifulsoup4==4.12.0
bs4==0.0.1
python-dateutil==2.8.2
requests==2.28.2
pandas==1.5.3
